+++ internal/line-number-plugin.js
@@ -7,6 +7,7 @@ const newLineRegex = /\r\n|\r|\n/
  * Options for the LineNumberPlugin. Currently nothing.
  * @typedef {object} Options
  * @property {boolean} [disableLineNumbers] - Toggle line numbers off.
+ * @property {number} [lineNumberStart=1] - Where to start counting from. Default is 1.
  */
 
 /**
@@ -17,7 +18,8 @@ export function LineNumberPlugin(
   options = {}
 ) {
 
-  const disableLineNumbers = options.disableLineNumbers
+  const disableLineNumbers = options.disableLineNumbers ?? false
+  const lineNumberStart = options.lineNumberStart ?? 1
 
   /**
    * @type {import('./prism-highlight.js').Hook}
@@ -33,7 +35,7 @@ export function LineNumberPlugin(
       }
 
       if (options.disableLineNumbers !== true) {
-        tokens.push(new Token("light-gutter-cell", (index + 1).toString()))
+        tokens.push(new Token("light-gutter-cell", (index + lineNumberStart).toString()))
       }
 
       tokens.push(new Token("light-line", ary))